requirements.txt
azure-core==1.33.0
azure-storage-blob==12.25.1
black==25.1.0
boto3==1.37.31
botocore==1.37.31
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coverage==7.8.0
cryptography==44.0.2
dnspython==2.7.0
google-cloud==0.34.0
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
jmespath==1.0.1
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
pymongo==4.11.3
pytest==8.3.5
pytest-cov==6.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
typing_extensions==4.13.1
urllib3==2.3.0
